+++ R/relabel.R
@@ -1,13 +1,16 @@
 #' Automatically relabel factor levels, collapse as necessary
 #'
-#' @param .f A factor.
-#' @param .fun A function that is applied to each level. Must accept one
-#'   character argument and return a character vector of the same length as its
-#'   input.
+#' @param f A factor.
+#' @param fun A bare or character function name or an actual function in
+#'   formula, quosure, or ordinary notation to be applied to each level. Must
+#'   accept one character argument and return a character vector of the same
+#'   length as its input.
 #' @param ... Additional arguments to `fun`.
 #' @export
 #' @examples
-#' fct_count(gss_cat$rincome)
+#'
+#' gss_cat$partyid %>% fct_count()
+#' gss_cat$partyid %>% fct_relabel(~gsub(",", ", ", .x)) %>% fct_count()
 #'
 #' convert_income <- function(x) {
 #'   regex <- "^(?:Lt |)[$]([0-9]+).*$"
@@ -17,15 +20,12 @@
 #'   x[is_range] <- paste0("Gt $", num_income)
 #'   x
 #' }
-#'
+#' fct_count(gss_cat$rincome)
 #' convert_income(levels(gss_cat$rincome))
-#'
 #' rincome2 <- fct_relabel(gss_cat$rincome, convert_income)
 #' fct_count(rincome2)
 fct_relabel <- function(.f, .fun, ...) {
-  if (!is.function(.fun)) {
-    stop("Expected function, got ", class(.fun)[[1L]], call. = FALSE)
-  }
+  .fun <- rlang::as_function(.fun)
 
   old_levels <- levels(.f)
   new_levels <- .fun(old_levels, ...)